api/requirements.txt
# Core dependencies
fastapi==0.109.0
sqlmodel==0.0.14
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25

# Pydantic and validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# Utils
python-dotenv==1.0.0
httpx==0.26.0